Worn out ritual change
The load function for Ninja Ritual with 'The Ritual is worn out' badge was setting the Ninja Ritual level to Infinity. There was no benefit gained from this as with the badge earned the code follows a path involving a constant 1e298 rather than the ritual level variable and it had the side effect of causing the ritual rift to cost infinity flux crystal.