Add a RewardRedacted call to RepeatableBoostUnlock
Hopefully this will break the loop where:

A boost such as Vault Key is locked and ASHF is inactive.
Because ASHF is inactive the first conditional group for shouldbuy is false, the second is also false because that is only true when always buy is set for Rob and the third conditional is not valid for Vault Key. Shouldbuy stays 0 / false and so Vault Key does not unlock.
Vault Key being locked means RewardLogicat will never reach the line where it falls back to RewardRedacted.

Also added an or for Between the Cracks being the same as Aleph One.

Putting as a pull request as someone else might have a more elegant solution.