If this fixes the bug, I quit.
(not really)

It's been a few weeks since I looked at this, for which I apologize. It mostly comes down to having a major project (which has been ongoing since January) due this coming Monday. And then I decided that it made sense to get back into coding regularly by learning Haskell.

Seriously though, if this fixes the bug, I'm going to be really mad.