Fix #1474
When adding t1i, I forgot to handle rifts initially, but everything looked OK, albeit harsh (since they always lead to OTC). It wasn't OK, but #1474 was the only way it could go wrong for now so it's all good. The other potential issue I spotted was that if your current timeline's max NP was larger than your max NP in OTC, you could use temporal rifts to get ahead in OTC. Though I'll admit it would be a funny easter egg for those who somehow were able to use it. Also, for simplicity, the code is now using `Math.ceil` instead of `Math.round`, so the chance of jumping to your max NP is a bit higher.